+++ src/main/java/br/com/fiap/challange/entity/StatusCombustivelEntity.java
@@ -13,8 +13,8 @@ public class StatusCombustivelEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer statusCombustivelId;
-	private Integer tipoCombustivelId;
+	private Long statusCombustivelId;
+	private Long tipoCombustivelId;
 	private String combustivelAdulterado;
 	
 	public StatusCombustivelEntity() {}
@@ -23,26 +23,26 @@ public class StatusCombustivelEntity {
 		this(model.getStatusCombustivelId(), model.getTipoCombustivelId(), model.getCombustivelAdulterado());
 	}
 	
-	public StatusCombustivelEntity(Integer statusCombustivelId, Integer tipoCombustivelId, String combustivelAdulterado) {
+	public StatusCombustivelEntity(Long statusCombustivelId, Long tipoCombustivelId, String combustivelAdulterado) {
 		super();
 		this.statusCombustivelId = statusCombustivelId;
 		this.tipoCombustivelId = tipoCombustivelId;
 		this.combustivelAdulterado = combustivelAdulterado;
 	}
 
-	public Integer getStatusCombustivelId() {
+	public Long getStatusCombustivelId() {
 		return statusCombustivelId;
 	}
 
-	public void setStatusCombustivelId(Integer statusCombustivelId) {
+	public void setStatusCombustivelId(Long statusCombustivelId) {
 		this.statusCombustivelId = statusCombustivelId;
 	}
 
-	public Integer getTipoCombustivelId() {
+	public Long getTipoCombustivelId() {
 		return tipoCombustivelId;
 	}
 
-	public void setTipoCombustivelId(Integer tipoCombustivelId) {
+	public void setTipoCombustivelId(Long tipoCombustivelId) {
 		this.tipoCombustivelId = tipoCombustivelId;
 	}
