+++ src/main/java/br/com/fiap/challange/model/AbastecimentoModel.java
@@ -7,8 +7,8 @@ public class AbastecimentoModel {
 	private Integer abastecimentoId;
 	private Date    abastecimentoDate;
 	private String  abastecimentoCoordenada;
-	private Integer statusCombustivelId;
-	private Integer tipoCombustivelId;
+	private Long statusCombustivelId;
+	private Long tipoCombustivelId;
 	private Integer veiculoId;
 	private Integer enderecoId;
 
@@ -36,19 +36,19 @@ public class AbastecimentoModel {
 		this.abastecimentoCoordenada = abastecimentoCoordenada;
 	}
 
-	public Integer getStatusCombustivelId() {
+	public Long getStatusCombustivelId() {
 		return statusCombustivelId;
 	}
 
-	public void setStatusCombustivelId(Integer statusCombustivelId) {
+	public void setStatusCombustivelId(Long statusCombustivelId) {
 		this.statusCombustivelId = statusCombustivelId;
 	}
 
-	public Integer getTipoCombustivelId() {
+	public Long getTipoCombustivelId() {
 		return tipoCombustivelId;
 	}
 
-	public void setTipoCombustivelId(Integer tipoCombustivelId) {
+	public void setTipoCombustivelId(Long tipoCombustivelId) {
 		this.tipoCombustivelId = tipoCombustivelId;
 	}
